+++ src/arithmetic.c
@@ -3,13 +3,13 @@
 
 #include "arithmetic.h"
 
-void add(emulatedCPU *cpu , uint16_t value) {
-    uint16_t result = (uint16_t) cpu->A + value;
+void add(emulatedCPU *cpu , uint8_t value) {
+    uint16_t result = (uint16_t) cpu->A + (uint16_t) value;
     updateAllFlags(cpu->cpuFlags,result);
     cpu->A = result & MAX_BYTE_VALUE_MASK;
 }
 
-void adc(emulatedCPU *cpu, uint16_t value) {
+void adc(emulatedCPU *cpu, uint8_t value) {
     add(cpu,value + cpu->cpuFlags->cy);
 }
 
@@ -21,7 +21,7 @@ void dad(emulatedCPU *cpu , uint16_t value) {
     cpu->L = result & MAX_BYTE_VALUE_MASK;
 }
 
-void sub(emulatedCPU *cpu, uint16_t value) {
+void sub(emulatedCPU *cpu, uint8_t value) {
     uint16_t result = (uint16_t) cpu->A - value;
     updateAllFlags(cpu->cpuFlags,result);
     cpu->A = result & MAX_BYTE_VALUE_MASK;
